+++ src/DepthBufferRender/DepthBufferRender.cpp
@@ -202,10 +202,10 @@ void Rasterize(vertexi * vtx, int vertices, int dzdx)
 	for (;;) {
 		int x1 = iceil(left_x);
 		if (x1 < (int)gDP.scissor.ulx)
-			x1 = gDP.scissor.ulx;
+			x1 = (int)gDP.scissor.ulx;
 		int width = iceil(right_x) - x1;
 		if (x1 + width >= (int)gDP.scissor.lrx)
-			width = gDP.scissor.lrx - x1 - 1;
+			width = (int)(gDP.scissor.lrx - x1 - 1);
 
 		if (width > 0 && y1 >= (int)gDP.scissor.uly) {
